fix(thumbnails): don't show screenshare thumb when it's not active
Fixs default layout.xml file when there is no screenshare, the default was show_presentation: true and show_screenshare: true and it showed the thumb even though screenshare wasn't shared.
Added validation to check if screenshare was shared before adding the thumb.
